+++ packages/web3-providers/src/types/Authorization.ts
@@ -1,0 +1,19 @@
+import type {
+    NonFungibleContractSpenderAuthorization,
+    FungibleTokenSpenderAuthorization,
+} from '@masknet/web3-shared-base'
+import type { SchemaType } from '@masknet/web3-shared-evm'
+
+export namespace AuthorizationAPI {
+    export interface Provider<ChainId> {
+        getApprovedNonFungibleContracts(
+            chainId: ChainId,
+            account: string,
+        ): Promise<Array<NonFungibleContractSpenderAuthorization<ChainId, SchemaType>>>
+
+        getApprovedFungibleTokenSpenders(
+            chainId: ChainId,
+            account: string,
+        ): Promise<Array<FungibleTokenSpenderAuthorization<ChainId, SchemaType>>>
+    }
+}
