+++ packages/web3-providers/src/types/GasOption.ts
@@ -1,0 +1,7 @@
+import type { GasOptionType } from '@masknet/web3-shared-base'
+
+export namespace GasOptionAPI {
+    export interface Provider<ChainId, GasOption> {
+        getGasOptions(chainId: ChainId): Promise<Record<GasOptionType, GasOption>>
+    }
+}
