+++ packages/web3-providers/src/types/RiskWarning.ts
@@ -1,0 +1,5 @@
+export namespace RiskWarningBaseAPI {
+    export interface Provider {
+        approve(address: string, pluginID?: string): Promise<void>
+    }
+}
