+++ packages/web3-providers/src/types/Token.ts
@@ -1,0 +1,10 @@
+export namespace TokenAPI {
+    export interface TokenInfo {
+        id: string
+        market_cap: string
+        price: string
+    }
+    export interface Provider {
+        getTokenInfo(tokenName: string): Promise<TokenInfo | undefined>
+    }
+}
