+++ packages/web3-providers/src/types/index.ts
@@ -1,0 +1,23 @@
+export * from './Authorization'
+export * from './Explorer'
+export * from './FungibleToken'
+export * from './GasOption'
+export * from './History'
+export * from './Instagram'
+export * from './Mask'
+export * from './NextID'
+export * from './NonFungibleToken'
+export * from './Price'
+export * from './RSS3'
+export * from './RiskWarning'
+export * from './Security'
+export * from './Storage'
+export * from './Token'
+export * from './Trending'
+export * from './Twitter'
+export * from './index'
+
+export enum NonFungibleMarketplace {
+    OpenSea = 'OpenSea',
+    LooksRare = 'LooksRare',
+}
